style(gateway): 优化注释和日志，完善代码风格细节
- 统一和丰富访问日志中过滤器类的注释说明，明确字段和常量含义
- 细化各配置类属性的注释，包括API版本、限流、跨域、IP访问控制等设置
- 完善HttpClient及连接提供者的注释，增强代码可读性
- 修正部分类型使用，如布尔类型替代整型状态标识
- 优化Redis对话记忆相关构造器及接口，增加API文档
- 修正列表分页响应中页码和页大小的数据类型
- 调整部分代码变量及方法命名，提升语义清晰
- 新增AI编码规范规则文件，规范@ApiVersion注解位置
- 增强网关指标和熔断器相关类注释，便于维护和监控
- 完善Token黑名单、IP访问控制、安全头等过滤器注释说明
- 统一SQL注入过滤规则注释及属性默认值说明
- 细化网关健康指标提供者字段说明，用于多维度监控
- 调整守护路径解码次数注释，提升安全防护文档说明

+++ ydsz-system/ydsz-system-web/src/main/java/com/njydsz/system/web/controller/ConfigApprovalController.java
@@ -62,8 +62,8 @@ public class ConfigApprovalController {
     PageResponse<List<ConfigApprovalVO>> pageResponse = new PageResponse<>();
     pageResponse.setData(items);
     pageResponse.setTotal((long) items.size());
-    pageResponse.setPageNum(query.getPageNum());
-    pageResponse.setPageSize(query.getPageSize());
+    pageResponse.setPageNum((long) query.getPageNum());
+    pageResponse.setPageSize((long) query.getPageSize());
     return YdszResponse.success(pageResponse);
   }
 
@@ -83,8 +83,8 @@ public class ConfigApprovalController {
     PageResponse<List<ConfigApprovalVO>> pageResponse = new PageResponse<>();
     pageResponse.setData(items);
     pageResponse.setTotal((long) items.size());
-    pageResponse.setPageNum(query.getPageNum());
-    pageResponse.setPageSize(query.getPageSize());
+    pageResponse.setPageNum((long) query.getPageNum());
+    pageResponse.setPageSize((long) query.getPageSize());
     return YdszResponse.success(pageResponse);
   }
 
@@ -103,8 +103,8 @@ public class ConfigApprovalController {
     PageResponse<List<ConfigApprovalVO>> pageResponse = new PageResponse<>();
     pageResponse.setData(items);
     pageResponse.setTotal((long) items.size());
-    pageResponse.setPageNum(query.getPageNum());
-    pageResponse.setPageSize(query.getPageSize());
+    pageResponse.setPageNum((long) query.getPageNum());
+    pageResponse.setPageSize((long) query.getPageSize());
     return YdszResponse.success(pageResponse);
   }
